removing anything not strictly necessary for author
