+++ packages/gamut-labs/src/index.ts
@@ -37,4 +37,4 @@ export * from './PageSection';
 export * from './ListSection';
 export * from './EmptySection';
 export * from './SocialMediaSharing';
-export * from './Tabs';
+export { Tabs as TabsExperimental, TabPanel } from './TabsExperimental';
