+++ packages/styleguide/stories/Molecules/SegmentedControlDeprecated/index.stories.mdx
@@ -1,0 +1,93 @@
+import {
+  HeadingDeprecated,
+  Tab,
+  TabList,
+  TabPanel,
+  Tabs,
+  Text,
+} from '@codecademy/gamut/src';
+import title from '@codecademy/macros/lib/title.macro';
+import { PropsTable } from '@codecademy/storybook-addon-variance';
+import { Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
+
+<Meta
+  title={title}
+  component={Tabs}
+  subcomponents={{ TabList, Tab, TabPanel }}
+  parameters={{
+    subtitle: `Use a series of tabbed areas for when you have multiple potential sections
+to show, but only one should be visible at a time.`,
+    source: 'gamut',
+    figmaId: 'fiLDZMjifzDmg70aYRUisW',
+    status: 'deprecated',
+  }}
+  args={{
+    renderAllPanels: false,
+    defaultActiveTabIndex: 2,
+    onChange: () => {},
+  }}
+/>
+
+<Canvas>
+  <Story name="Tabs">
+    {(args) => (
+      <Tabs {...args}>
+        <TabList>
+          <Tab>Tab 1</Tab>
+          <Tab>Tab 2</Tab>
+          <Tab>Tab 3</Tab>
+        </TabList>
+        <TabPanel>
+          <HeadingDeprecated as="h2">Welcome to Tab 1</HeadingDeprecated>
+          <Text>Hi there! I&apos;m the contents inside Tab 1. Yippee!</Text>
+        </TabPanel>
+        <TabPanel>
+          <HeadingDeprecated as="h2">Welcome to Tab 2</HeadingDeprecated>
+          <Text>Hi there! I&apos;m the contents inside Tab 2. Yippee!</Text>
+        </TabPanel>
+        <TabPanel>
+          <HeadingDeprecated as="h2">Welcome to Tab 3</HeadingDeprecated>
+          <Text>Hi there! I&apos;m the contents inside Tab 3. Yippee!</Text>
+        </TabPanel>
+      </Tabs>
+    )}
+  </Story>
+</Canvas>
+
+<PropsTable story="Tabs" />
+
+## Usage
+
+Here's an example usage of any uncontrolled TabList
+
+## Controlled
+
+When `activeTabIndex` is specified, the component switches to
+being a controlled component. It won't change its own tab state, but will
+instead notify of a request to change using `onChange`.
+
+<Canvas>
+  <Story name="Controlled" args={{ activeTabIndex: 0 }}>
+    {(args) => (
+      <Tabs {...args}>
+        <TabList>
+          <Tab>Tab 1</Tab>
+          <Tab>Tab 2</Tab>
+          <Tab>Tab 3</Tab>
+        </TabList>
+        <TabPanel>
+          <HeadingDeprecated as="h2">Welcome to Tab 2</HeadingDeprecated>
+          <Text>Hi there! I&apos;m the contents inside Tab 2. Yippee!</Text>
+        </TabPanel>
+        <TabPanel>
+          <HeadingDeprecated as="h2">Welcome to Tab 1</HeadingDeprecated>
+          <Text>Hi there! I&apos;m the contents inside Tab 1. Yippee!</Text>
+        </TabPanel>
+        <TabPanel>
+          <HeadingDeprecated as="h2">Welcome to Tab 3</HeadingDeprecated>
+          <Text>Hi there! I&apos;m the contents inside Tab 3. Yippee!</Text>
+        </TabPanel>
+      </Tabs>
+    )}
+  </Story>
+</Canvas>
