--- packages/styleguide/stories/Molecules/Tabs/index.stories.mdx
@@ -1,149 +1,0 @@
-import { Text } from '@codecademy/gamut';
-import { TabPanel, Tabs } from '@codecademy/gamut-labs';
-import title from '@codecademy/macros/lib/title.macro';
-import { PropsTable } from '@codecademy/storybook-addon-variance';
-import { Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
-import { useState } from '@storybook/addons';
-
-import ExampleCard from './ExampleCard';
-
-<Meta
-  title={title}
-  component={Tabs}
-  subcomponents={{ TabPanel }}
-  parameters={{
-    subtitle: `Use a series of tabbed areas for when you have multiple potential sections
-to show, but only one should be visible at a time.`,
-    source: 'gamut-labs',
-    figmaId: '',
-    status: 'updating',
-  }}
-  args={{
-    activeTabIndex: 0,
-  }}
-/>
-
-<Canvas>
-  <Story name="Tabs">
-    {(args) => (
-      <Tabs {...args}>
-        <TabPanel title="On soccer">
-          <ExampleCard
-            src="https://static.wikia.nocookie.net/kingofthehill/images/c/c7/Bobby.png"
-            quote="You know Dad, the kids playing soccer don't have any lumps or bruises and Coach Lucas gives them orange slices at half time."
-            speaker="Bobby Hill"
-          />
-        </TabPanel>
-        <TabPanel title="On vampires">
-          <ExampleCard
-            src="https://static.wikia.nocookie.net/adventuretimewithfinnandjake/images/8/80/S4e25_Marceline_happy.png"
-            speaker="Marcline the Vampire Queen"
-            quote="Vampires can't beat ghosts. It's like a rock-paper-scissors thing."
-          />
-        </TabPanel>
-        <TabPanel title="On selflessness">
-          <ExampleCard
-            src="https://static.wikia.nocookie.net/dragonball/images/a/a9/Piccolo10YearsLater.png"
-            quote="Sometimes, we have to look beyond what we want and do what's best."
-            speaker="Piccolo"
-          />
-        </TabPanel>
-      </Tabs>
-    )}
-  </Story>
-</Canvas>
-
-<PropsTable story="Tabs" />
-
-## Tabs with defined width
-
-<Canvas>
-  <Story name="Tabs with defined width">
-    <Tabs tabSize={300}>
-      <TabPanel title="On toughness">
-        <ExampleCard
-          src="https://static.wikia.nocookie.net/starwars/images/7/71/Ahsoka_The_Padawan_Path.jpg"
-          speaker="Ahsoka Tano"
-          quote="You don’t have to look tough to be tough."
-        />
-      </TabPanel>
-      <TabPanel title="On summer rains and their predictability">
-        <ExampleCard
-          src="https://static.wikia.nocookie.net/edwikia/images/0/0a/DDPhone.JPG"
-          quote="Summer rains, you can never predict them."
-          speaker="Edd"
-        />
-      </TabPanel>
-    </Tabs>
-  </Story>
-</Canvas>
-
-## Tabs with custom active color
-
-<Canvas>
-  <Story name="Tabs with custom active color">
-    <Tabs activeTabColor="red-500">
-      <TabPanel title="On possibilities">
-        <ExampleCard
-          src="https://static.wikia.nocookie.net/steven-universe/images/c/cf/Garnet_No_Visor2.png"
-          speaker="Garnet"
-          quote="There are millions of possibilities for the future, but it's up to you to choose which becomes reality. Please understand. You choose your own future."
-        />
-      </TabPanel>
-      <TabPanel title="On stoicism">
-        <ExampleCard
-          src="https://static.wikia.nocookie.net/samuraijack/images/b/b3/Jack%27s_nirvana.png"
-          quote="..."
-          speaker="Jack"
-        />
-      </TabPanel>
-    </Tabs>
-  </Story>
-</Canvas>
-
-# TabPanel
-
-## Disabled TabPanel
-
-<Canvas>
-  <Story name="Disabled TabPanel">
-    <Tabs>
-      <TabPanel title="On friendship">
-        <ExampleCard
-          src="https://static.wikia.nocookie.net/spongebob/images/9/9a/Prehibernation_Week_044.png"
-          quote="If you want to be my friend, just be yourself!"
-          speaker="Sandy Cheeks"
-        />
-      </TabPanel>
-      <TabPanel title="On mystery" tabDisabled />
-    </Tabs>
-  </Story>
-</Canvas>
-
-## TabPanel with onClick
-
-<Canvas>
-  <Story name="TabPanel with onClick">
-    {() => {
-      const [points, setPoints] = useState(0);
-      return (
-        <>
-          <Tabs>
-            <TabPanel
-              title="On perseverance"
-              onTabClick={() => setPoints(points + 1)}
-            >
-              <ExampleCard
-                src="https://static.wikia.nocookie.net/kurokonobasuke/images/7/79/LG_Kuroko.png"
-                quote="The possibility of victory is only 0% when all the players give up. No matter how impossible it may seem, I refuse to make it 0% myself."
-                speaker="Tetsuya Kuroko"
-              />
-            </TabPanel>
-          </Tabs>
-          <br />
-          <Text>Kuroko's score: {points}</Text>
-        </>
-      );
-    }}
-  </Story>
-</Canvas>
