+++ packages/styleguide/stories/Molecules/TabsExperimental/index.stories.mdx
@@ -1,0 +1,49 @@
+import { TabPanel, TabsExperimental } from '@codecademy/gamut-labs';
+import title from '@codecademy/macros/lib/title.macro';
+import { PropsTable } from '@codecademy/storybook-addon-variance';
+import { Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
+
+import ExampleCard from './ExampleCard';
+
+<Meta
+  title={title}
+  component={TabsExperimental}
+  subcomponents={{ TabPanel }}
+  parameters={{
+    subtitle: `Use a series of tabbed areas for when you have multiple potential sections
+to show, but only one should be visible at a time.`,
+    source: 'gamut-labs',
+    figmaId: '',
+    status: 'updating',
+  }}
+/>
+
+<Canvas>
+  <Story name="Tabs (Experimental)">
+    <TabsExperimental>
+      <TabPanel title="On soccer">
+        <ExampleCard
+          src="https://static.wikia.nocookie.net/kingofthehill/images/c/c7/Bobby.png"
+          quote="You know Dad, the kids playing soccer don't have any lumps or bruises and Coach Lucas gives them orange slices at half time."
+          speaker="Bobby Hill"
+        />
+      </TabPanel>
+      <TabPanel title="On vampires, ghosts, and other spooky creatures">
+        <ExampleCard
+          src="https://static.wikia.nocookie.net/adventuretimewithfinnandjake/images/8/80/S4e25_Marceline_happy.png"
+          speaker="Marcline the Vampire Queen"
+          quote="Vampires can't beat ghosts. It's like a rock-paper-scissors thing."
+        />
+      </TabPanel>
+      <TabPanel title="On selflessness">
+        <ExampleCard
+          src="https://static.wikia.nocookie.net/dragonball/images/a/a9/Piccolo10YearsLater.png"
+          quote="Sometimes, we have to look beyond what we want and do what's best."
+          speaker="Piccolo"
+        />
+      </TabPanel>
+    </TabsExperimental>
+  </Story>
+</Canvas>
+
+<PropsTable story="Tabs (Experimental)" />
